yt_comments_scraper/requirements.txt
beautifulsoup4==4.12.2
pandas==2.1.4
python_dateutil==2.8.2
selenium==4.16.0
